+++ Services/FlipTrackingService.cs
@@ -103,7 +103,7 @@ namespace Coflnet.Sky.Commands
             var res = await flipTracking.TrackerFlipAuctionIdPostAsync(flip.Auction.Uuid, new FlipTracker.Client.Model.Flip()
             {
                 FinderType = (FlipTracker.Client.Model.FinderType?)flip.Finder,
-                TargetPrice = flip.TargetPrice,
+                TargetPrice = (int)flip.TargetPrice,
                 Timestamp = foundAt,
                 AuctionId = flip.UId
             }).ConfigureAwait(false);
